+++ CHANGELOG.md
@@ -7,8 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Security`
 
-
-## [Unreleased] 
+## [Unreleased]
 
 ### Added
 
@@ -21,6 +20,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 ### Changed
 
 - SC-6080 - move REQUEST_TIMEOUT from globals to Configuration
+- Dependencies: querystring replaced by qs
 - SC-6060 - Updated error handling
 
 ### Fixed